Fix CREATE_NEW_PURCHASE Procedure and Utilize in Code The sp now works and is used in the code to make purchases. All that is needed to make a purchase is to call this procedure; it will adjust everything else and create the necessary linked entries.
Note - this change causes an issue as the sp modifies the CUSTOMER object when it updates the balance. Therefore, when we later try to make changes in the code to our older reference to the same customer object, it throws an exception. See if there's a way to refresh the object to fix this issue.